Moved admin-only helpers into admin side
ConsultationsHelper#consultation_opening_phrase and consultation_closing_phrase were only being used in the admin suite.
They are now included in an admin helper.

Also updated the language to be 'Opens at' instead of 'Opens on'
